+++ sdk/monitor/azure-monitor-opentelemetry-exporter/CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Features Added
 - Customer Facing Statsbeat: Added remaining drop codes to base
   ([#42382](https://github.com/Azure/azure-sdk-for-python/pull/42382))
+- Refactored the put methods in storage.py for LocalFileBlob and LocalFileStorage
+  ([#42502](https://github.com/Azure/azure-sdk-for-python/pull/42502))
 
 ### Breaking Changes